Refactor: remove deprecated imports from storageService in config and dspace-client service files
