+++ src/cli/commands/config.ts
@@ -1,5 +1,4 @@
-import { Config } from '../services/config.service'
-import { storageService } from '../services/storage.service'
+import { storageService, Config } from '../services/storage.service'
 import { dspaceClient } from '../services/dspace-client.service'
 
 export const configCommands = {
